+++ mlir/lib/Dialect/AMX/Analysis/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_mlir_dialect_library(MLIRAMXAnalysis
+  AMXBindingAnalysis.cpp
+
+  DEPENDS
+  MLIRAMXConversionsIncGen
+
+  LINK_LIBS PUBLIC
+  MLIRAMXDialect
+  MLIRIR
+  )
